# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a dynamic Headless CMS API built with Go Fiber and MongoDB.",
        "title": "GoHeadless CMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/content/{collSlug}": {
            "get": {
                "description": "Fetch all dynamic content records for the given collection slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "List all records in a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "collSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Upsert dynamic content into a collection. Validates against collection schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Create a new record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "collSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dynamic Record Data",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/content/{collSlug}/{id}": {
            "get": {
                "description": "Retrieve a specific dynamic record by its unique MongoDB ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "collSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record ID (Hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Modify a dynamic record. Validates update against collection schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Update record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "collSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record ID (Hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Record Data",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "description": "Remove a specific dynamic record from a collection",
                "tags": [
                    "content"
                ],
                "summary": "Delete record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection Slug",
                        "name": "collSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Record ID (Hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/upload": {
            "put": {
                "description": "Deletes the old file and replaces it with a new one.",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Replace an uploaded file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The path of the file to replace",
                        "name": "oldPath",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "The new file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the new file path",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during processing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Handles file upload, converts images to WebP if applicable, and saves it to the server's local storage.",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the file path",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file upload request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during processing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a file from the server's local storage based on its path.",
                "tags": [
                    "Upload"
                ],
                "summary": "Delete an uploaded file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The path of the file to delete (e.g. /uploads/file_name.ext)",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid path or file not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during file deletion",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}